Improve configuration file checks in entrypoint script

+++ games/azerothcore/entrypoint.sh
@@ -4,15 +4,11 @@ set -e
 
 cd /home/container
 
-if [ -f "/home/container/env/dist/etc/authserver.conf.dist" ]; then
+if [ -f "/home/container/env/dist/etc/authserver.conf.dist" ] && [ ! -f "/home/container/env/dist/etc/authserver.conf" ]; then
     cp /home/container/env/dist/etc/authserver.conf.dist /home/container/env/dist/etc/authserver.conf
-else
-    echo "No authserver.conf.dist file found, this was not expected."
 fi
-if [ -f "/home/container/env/dist/etc/worldserver.conf.dist" ]; then
+if [ -f "/home/container/env/dist/etc/worldserver.conf.dist" ] && [ ! -f "/home/container/env/dist/etc/worldserver.conf" ]; then
     cp /home/container/env/dist/etc/worldserver.conf.dist /home/container/env/dist/etc/worldserver.conf
-else
-    echo "No worldserver.conf.dist file found, this was not expected."
 fi
 
 # Fix permissions in case Pterodactyl mounted files
